+++ crates/dav/src/calendar/get.rs
@@ -124,7 +124,7 @@ impl CalendarGetRequestHandler for Server {
         if !is_head {
             Ok(response.with_binary_body(ical))
         } else {
-            Ok(response.with_content_length(ical.len()))
+            Ok(response.with_content_length(ical.len() as u64))
         }
     }
 }
