+++ crates/jmap/src/blob/mod.rs
@@ -19,5 +19,5 @@ pub struct UploadResponse {
     blob_id: BlobId,
     #[serde(rename(serialize = "type"))]
     c_type: String,
-    size: usize,
+    size: u64,
 }
